authorization module + bugfixes & improvements

+++ src/Modules/Machine/Routes/api.php
@@ -11,9 +11,11 @@
 |
 */
 
-Route::get('/{id}', 'MachineController@show');
-Route::patch('/{id}', 'MachineController@update');
-Route::delete('/{id}', 'MachineController@destroy');
+use Modules\Authorization\Entities\Permission;
 
-Route::post('/', 'MachineController@store');
-Route::get('/', 'MachineController@index');
+Route::get('/{id}', 'MachineController@show')->middleware(['permission:'.Permission::SHOW_MACHINE]);
+Route::patch('/{id}', 'MachineController@update')->middleware(['permission:'.Permission::UPDATE_MACHINE]);
+Route::delete('/{id}', 'MachineController@destroy')->middleware(['permission:'.Permission::DELETE_MACHINE]);
+
+Route::post('/', 'MachineController@store')->middleware(['permission:'.Permission::CREATE_MACHINE]);
+Route::get('/', 'MachineController@index')->middleware(['permission:'.Permission::SHOW_MACHINE]);
